feat: Add delete button confirmation and preliminary routing
Update the book-viewer.html "delete" button using a JS script that
asks the user if they really want to delete the book, and then routes
them back to the book's genre's main page. Still need to implement
functionality of button to actually delete the book by re-routing the
browser to a uri that deletes the book, may need to figure out how
to make the redirect use the "post" method.