refactor: Upgrade JS script regarding image search call
Now the script waits until the entire document has loaded, and then
updates the src to the URL for the custom image searched book cover.
This removes the "on.('load', ...)" JQuery operation that is known
to be buggy for this use case -> (https://api.jquery.com/load-event/),
so the script does not wait until just the placeholder .gif image has
loaded.