feat: Add "Add New Genre" and "Add New Book" buttons to genreIndex.html and genre-list.html pages
Note: Also changes the Login button on book-viewer-logged-out.html
page from a static button to one that becomes a Login/Logout button
dependent on the viewer's login status (e.g. a different user from the
one who added a book may be viewing it, and not able to edit it but
able to log out now if they so choose!).