feat: Enable all pages and functions work on website
Note: Started navigating by book titles as opposed to book id's
for identifying the books, which causes minor issues in loading the
correct book for duplicates when a new book with the same title is
added, for seeing the book description edit update for the book you
are editing if it is a duplicate, etc. Due to older SQL formatting
for PostgreSQL vs. SQLAlchemy, could look into work-arounds where
the duplicate books are referenced by their book ids and not just
their titles, but the problem is minor overall.